WIP: MetricSet, addMetric
